Feat ajusto con microservicios relacionados

+++ src/main/java/cl/catastrofescl/citizen/dto/response/DonationItemResponse.java
@@ -5,6 +5,6 @@ import java.util.UUID;
 public record DonationItemResponse(
         UUID id,
         UUID itemId,
-        Integer cantidad
+        Long cantidad
 ) {
 }
